+++ com.unity.render-pipelines.high-definition/Editor/VFXGraph/VFXHDRPSubOutput.cs
@@ -16,6 +16,9 @@ namespace UnityEditor.VFX
         [VFXSetting(VFXSettingAttribute.VisibleFlags.InInspector), Header("HDRP"), Tooltip("Specifies when in the render queue particles are drawn. This is useful for drawing particles behind refractive surfaces like frosted glass, for performance gains by rendering them in low resolution, or to draw particles after post processing so they are not affected by effects such as Depth of Field.")]
         public TransparentRenderQueue transparentRenderQueue = TransparentRenderQueue.Default;
 
+
+        
+
         // Caps
         public override bool supportsExposure { get { return true; } } 
         public override bool supportsMotionVector