+++ com.unity.visualeffectgraph/Editor/Models/Operators/Implementations/SquaredDistance.cs
@@ -30,11 +30,11 @@ namespace UnityEditor.VFX.Operator
             }
         }
 
-        protected override sealed VFXPropertyAttribute[] expectedOutputAttributes
+        protected override sealed VFXPropertyAttributes expectedOutputAttributes
         {
             get
             {
-                return VFXPropertyAttribute.Create(new TooltipAttribute("The squared distance between a and b."));
+                return new VFXPropertyAttributes(new TooltipAttribute("The squared distance between a and b."));
             }
         }